making our website less static.
We added a header and a footer to our web pages - this is also a navigation bar (currently not working).
We added a lovely picture of phoebe. This is coded so that it will work on a real webpage.
Css file is updated so that everything is looking lovely and all set out correctly/coloured correctly.